Import 2.1.54

+++ drivers/block/rd.c
@@ -447,7 +447,7 @@ __initfunc(static void rd_load_image(kdev_t device,int offset))
  	struct inode inode, out_inode;
 	struct file infile, outfile;
 	struct dentry in_dentry, out_dentry;
-	unsigned short fs;
+	unsigned long fs;
 	kdev_t ram_device;
 	int nblocks, i;
 	char *buf;
